Added hack to work around bug in Piston.
Piston can't handle content type headers that also include character set data.
To resolve this, I added a decorator that strips off the char set data. As a
bonus, the decorator uses the char set data to inform the encoding of the
payload (another bug in Piston prevents it from benefitting from this, but
it's still nice).